[project @ 1518] Added a build-db script to do the createdb, load, indexes, and foreign keys.  I renamed associated scripts to what I think are easier names.
Also made some /bin/sh to /bin/bash changes that should help
portability to other os's.  Linux /bin/sh usually defaults to bash but
when other os's don't default to bash we have syntax problems when the
script has bash-isms.

Also fixed the path order in configure.ac so that default paths don't
come first when a non-default postgresql installation is specified.

Also fixed the location of AC_CONFIG_FILES for scripts that need to be
created when --with-postgresql is not specified.