A bunch of changes that came out of testing against the driver.
Removed unecessary semicolons (;) from the end of SQL statements.

Correctly set variables to long, which were int.  Symptoms were that the
values passed differed when they were bigger than the maximum integer.

Fixed the logic of generating multi-dimensional arrays in more than one
location.

Changed the handling of padded CHAR to unpadded VARCHAR.

Corrected a query in the trade lookup frame.

Updated the logic to handle BOOLEAN in the database while the stored
function is still using SMALLINT as IN and OUT parameters.

Corrected the calculation of the values[] that hold the data to return
in several frames.

Improved the elog() message when unexpected event or errors occur in the
functions.