Correct several things in the data maintenance frame.
In any unexpected event, return 1, which indicates and error.

Fixed the input dumping functions to use longs instead of ints for the
appropriate input data.

Fixed a query for the "customer" portion to use the correct c_id, as
opposed to acct_id.

Added a FIXME note for a previous NULL pointer seg fault.