Several fixes to the C stored functions.
Added double quotes to items in the string arrays to explicitly
identify a string.  Some string had commas in them so data was being
split incorrectly.

Fixed the calculation of asset_total in the customer position frame.
Sometimes a NULL value is returned so it needs to be handled properly.

Fixed a couple of places where the length of various arrays were being
defined incorrectly.  Was using the incorrectly macro to define the
data size.

Removed code to fail the frame when no data was returned from SELECT
statements in the trade update frame because 0 rows can be returned.