unique identifier fix
The kit was incorrectly taking a unique identifier from the command line
and assigning to all the EGen classes, except for the thread running the
data maintenance transactions.  The unique idetifier is not set by the
thread's pthread id, and the programs now optionally take a random
generator seed from the command line, which invalidates test runs.  This
is allowed behavior if so desired.