fix compiler warnings
Previously not using building with -Wall due to a enormous amounts of
messages spewed from the kit, most likely due to a combination of the
version of the C++ compile EGen was written for and the C++ compiler
used to develop the kit.  The warnings are under control now, if not
completely gone.  These are the fixes to the warnings for the DBT5
portion of the kit.  EGen warnings still remain.