postgresql stored function fixes
The length defined for status (STATUS_LEN) was too small so proper
status couldn't be returned.  Updated from 4 to 5 to handle negative
four digit codes.  e.g. "-1234"

The Trade Result stored function corrected not to fail the frame if zero
rows are returned in the first query, and to continue appropriately.

The Trade Status stored function corrected to return a -911 error code
if 50 rows of data are not processed.