security detail transaction fix
In postgresql the ni_item is bytea of unlimited size.  Well, max size of
a bytea columm which is bigger than the 100,000 character limit capped
by EGen.  As a bytea column, depending of the character encoding, each
character can be escaped by 5 characters, meaning the total string
returned can be blown up 5 times.

The stored function has its buffer upped 5 times for ni_item and strncpy
is used in more places to prevent buffer overflows and the database
backend from crashing.

The driver size of the kit truncates the string returned back to 100,000
characters.  This portion may still have to be fix addressed later.