postgresql empty result set error handling
Fixed application logic on the driver size regarding 0 result sets
returned by the database stored function.  The workload is designed to
always return 1 row of data for each transaction frame.  If a frame
returns no data than there is some problem with the test.  In this even,
the code attempts to handle it gracefully by logging an error and
marking the transaction status to rollback.