security detail fix
This bug manifested when the cardinality of the customers was increased.
The PostgreSQL stored function was returning the s_num_out as a 32-bit
integer, which caused errors when the scale of the customers increased
the possible values beyond that range.  The stored function definition
was fixed and the transaction code also has to be fixed to handle values
that would overrun a C int.