fix buffer overruns in pg stored functions
Fix a bug in the data maintenance transaction where a table name may
overrun the length of the buffer that was supposed to hold it.

Also identified other areas where even though strnlen was preventing a
copy from overrunning a buffer, it did not guarantee that the resulting
string was null terminated.  This resulted in a string being read beyond
its buffer.