update for egen 1.12.0 changes
Remove all references to status in the PostgreSQL C stored functions and
relating parts of the kit.

Use CBaseTxnErr::EXPECTED_ROLLBACK instead of CBaseTxnErr::ROLLBACK
since it doesn't exist anymore.

MarketFeed zz_padding turned into zz_padding1 and zz_padding2.

Fixed the #DEBUG code in the transaction code.  Print the id of the
current thread to differentiate between running threads in the logs.
Would prefer locks since some of the cout messages still clobber each
other at times, but sharing locks will take a little more work.

The inline escape_me() function for the PostgreSQL C stored functions
stopped working.  Manually inlined the code in the 3 places it was being
used.

Removed the RETURNING clause for a SQL statement in the PostgreSQL
Market-Feed frame since the value wasn't being used.

Fixed the format of the array output of several fields in the PostgreSQL
C stored function for Trade-Lookup.

Fixed the PostgreSQL Trade-Lookup Frame 4 to not execute a SQL statement
TLF4_2 if the previous query doesn't retrieve any data for it.

Fixed PostgreSQL Trade-Result Frame 2 to not fail the frame when no data
is returned from one of the queries.  This behavior is supposed to be
ok.

Fixed PostgreSQL Trade-Update Frame 1 to not fail the frame when no data
is returned from one of the queries.  This behavior is supposed to be
ok.

Fixed the format of the array output of several fields in the PostgreSQL
C stored function for Trade-Update Frame 1.