DBT-5 fixes
1) src/BrokerageHouse/BrokerageHouse.cpp, src/include/BrokerageHouse.h
Each thread should assign individual buffer area (bvOutput etc.) for storing
transaction result. As to current implementation, all threads use the same
buffer area located in the instance of CBrokerageHouse object.

I believe it is unnecessary to output error message for RunTradeOrder when
intentional roll back is specified.

2)src/include/CommonStructs.h, src/include/DBT5Consts.h
Transaction id number is better to be consistent with the definition of
egen/inc/CETxnMixGenerator.h. As scripts to summarize benchmark execution
result (dbt5-post-process) uses the definition of
egen/inc/CETxnMixGenerator.h, current implementation fails to output result
correctly.

3)src/include/MEESUT.h, src/interfaces/MEESUT.cpp
Multiple threads can use the socket of the MEESUT object simultaneously, and
thus, mutual exclusion is necessary to use the socket. Although this mutual
exclusion can be a bottleneck in the benchmark execution, it is introduced,
since some radical changes seem to be necessary to address this issue.

4)src/transactions/pgsql/DBConnection.cpp
4-1)MarketFeedFrame shoud set 'num_updated' to the output frame specified by
the 'pOut' pointer.

4-2)TradeLookupFrame4() stored procedure should return 'num_trades_found' to
confirm the TPC-E specification..

5)storedproc/pgsql/c
5-1)Common problems
The SPI_processed value changes by the execution of SPI_execute_plan(), and
thus the value has to be saved in another variable when the value is used in
the exit condition of a loop in which SPI_execute_plan() is executed.

5-2)market_feed.c
'#include <access/tupmacs.h>' is necessary to use it against PostgreSQL
varsion 9.3.

The forth parameter for MFF1_2 (args[3]) differed from the TPC-E
specification.

5-3)trade_cleanup.c
The position of the DELETE statement differed from the TPC-E specification.

5-4)trade_lookup.c, trade_lookup.sql_in
Variables start_trade_dts_ts and end_trade_dts_ts have to be used for
TLF3_1, not start_trade_dts and end_trade_dts.

'num_trades_found' is included in the return values of TradeLookupFrame4().

The 'num_found_count = SPI_processed;' statement is moved into the if
succeeded case. (It seems natural for me.)

5-5)trade_update.c
'num_updated' has bo be counted up.

Dummy data are returned as cash_transaction_amount, cash_transaction_dts,
and cash_transaction_name when is_cash_str[0] != 't'. There seems no easy
way to return NULL value as an element of an array data.

Some simple mistakes are found in the 'for (i = 0; i < num_found; i++) {'
loop in TradeUpdateFrame3().

6)README-POSTGRESQL
I believe '-t' option is necessary for 'dbt5-pgsql-build-db'.